refactor: Update gelflogger and introduce NewZeroLogger
Implemented a baseLogProcessor in gelflogger's NewLogger for flexible
log field processing. This change improves modularity by shifting the
task of field processing from the exclusive scope of
`formatGELFMessage`, allowing each log entry to be processed
individually within the `Log` function. A redundant test case was also
removed for clarity. Lastly, the `NewZeroLogger` function was added in
`zerologger.go` to create a zerolog logger with a GelfWriter.